Pull components out of graph page
